proto_3d_converter/requirements.txt
torch
transformers
diffusers
rembg
opencv-python
numpy
trimesh
pyrender
matplotlib
Pillow